Land Recent QUIC changes.
Make the FEC group optional by adding a flag to the private headers.

Merge internal change: 46979143

Merging changes from chromium CL - 15385004

Merge internal change: 46949614

Removing debug logging from RecordPacketReceived.  Seems redundant to
log both when we actually receive and when we record it.

Merge internal change: 46934210

Logging crypto handshake as a DVLOG rather than DLOG as it hasn't
recently been needed to debug test failures.

Merge internal change: 46932247

Changing the quic test client to simply not return a stream if not connected.

This will hopefully turn server test check-failures into server test
failures.

Merge internal change: 46932163

QUIC: redo server nonces.

Previously, in order to cope with strike-register failures and client
clock-sync issues, the server could issue a server nonce to a client. This
meant that the server had to remember rejected handshakes so that the server
nonce could be matched up. With this change, QUIC servers no longer need to
keep track of rejected handshakes.

Instead of issuing and remembering nonces, a server will now encrypt them and
forget about them. When a server nonce is used to establish freshness for a
connection, it will be stored in a per-GFE strike-register. (This
strike-register is separate from the one used to process client nonces.)

Merge internal change: 46889484

Remove FEC_ENTROPY_FLAG from private flags. Now, FEC packet's entropy
flag contain the xor of entropies of the protected packets.

Merge internal change: 46889094

Limit the number of times we'll fast-retransmit a given packet using taildrop.

Merge internal change: 46754530

Added CommonCertSetsQUIC to anonymous namespace.


QUIC: cleanups round two.

 * Make CommonCertSetsQUIC a Singleton to save on every Config having its own
   copy.
 * Rework server config expiry: previously it caused an error at client hello
   send time. Now it will cause an error at REJ processing time but, if the
   config expired after we cached it, we will act as if we didn't have a cached
   server config.
 * Invalidate the server config cache in the event of a client hello sending
   failure. This will prevent a bad server config from being cached and poisoning
   connection attempts for the lifetime of the cache.
 * Fix a bug in the test code which failed to parse hex chunks in debugging
   messages correctly. (Thanks to wtc for noticing.)

Merge internal change: 46742937

Merging changes from chromium - CL 15074007

Merge internal change: 46710932

Fix a bug in QuicSession's header compression behavior which could lead
to infinite loops.

Merge internal change: 46694681

Getting 5% our CPU usage back by not calculating SentBandwidth for the
tcp congestion control algorithm.

Added a TODO to improve that function since it's pretty abysmal: the
ToLargerUnits and Subtract overhead alone accounted for 4.5% of the cpu
in initial loadtest runs.

Merge internal change: 46608880

Adding support for truncated guids in QuicFramer.

Merge internal change: 46575819

using our latched write_blocked status to spare us useless system calls.

Merge internal change: 46573462

Fixing some crashing issues in the QUIC loadtest, where if a client ever
disconnects it never recovers, either crashing trying to create a stream
or crashing waiting for a response on a non-existant stream.

I'm not sure if we have the same problem for the http/https simple clients
but we definitely do for QUIC.

Merge internal change: 46562890

Merging changes from chromium - CL 14614006

Merge internal change: 46460427

Merging cleanup changes from chromium CL - 14651009

Merge internal change: 46457093

Fixing a test framework bug for quic: we were munging headers to do
https:// for insecure quic resulting in a 404 in the http-only service
map. Then disalbing the test since we don't advertise secure SPDY on
insecure QUIC.

Merge internal change: 46408400

Move QuicConfig from ssl_global_data to quic_dispatcher.cc. Initialize
using values from QuicConfigProto and use the max_time_before_crypto_handshake
to set the overall connection timeout before crypto handshake finishes.

Merge internal change: 46400649

QUIC: implement ChannelIDs.

We'll need this for HTTPS.

Merge internal change: 46396357

Deleted usage of scoped_ptr_openssl. Added TODO comments for porting
ChannelIDSigner and Verifier.

R=rch@chromium.org

Review URL: https://chromiumcodereview.appspot.com/15937012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@203220 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b0643107853022d9925317352099d525b9eaa359

+++ net/quic/quic_stream_sequencer.h
@@ -59,11 +59,11 @@ class NET_EXPORT_PRIVATE QuicStreamSequencer {
 
   // Fills in up to iov_len iovecs with the next readable regions.  Returns the
   // number of iovs used.  Non-destructive of the underlying data.
-  int GetReadableRegions(iovec* iov, int iov_len);
+  int GetReadableRegions(iovec* iov, size_t iov_len);
 
   // Copies the data into the iov_len buffers provided.  Returns the number of
   // bytes read.  Any buffered data no longer in use will be released.
-  int Readv(const struct iovec* iov, int iov_len);
+  int Readv(const struct iovec* iov, size_t iov_len);
 
   // Consumes |num_bytes| data.  Used in conjunction with |GetReadableRegions|
   // to do zero-copy reads.
